realizado rotas
