--- quinzena9/labex/src/components/DetailsTravel.js
@@ -1,7 +1,0 @@
-export function DetailsTravel() {
-    return (
-        <div>
-            Detalhes da Viagem
-        </div>
-    )
-}
