+++ quinzena9/labex/src/pages/DetailsTravel.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import { useHistory } from "react-router-dom"
+
+export function DetailsTravel() {
+    const history = useHistory()
+
+    const goToHome = () => {
+        history.push("/")
+    }
+    return (
+        <div>
+            <button onClick={goToHome}>voltar para home</button>
+            Detalhes da Viagem
+        </div>
+    )
+}
