+++ quinzena9/labex/src/pages/Subscription.js
@@ -1,0 +1,38 @@
+import React from 'react'
+import { useHistory } from 'react-router-dom'
+import styled from 'styled-components'
+
+const RegisterContainer = styled.div`
+  border-radius: 50px;
+  background-color: white;
+  width: 250px;
+  height: 200px;
+  margin: 50px 10px;
+`
+
+const ContainerInputs = styled.div`
+  text-align: center;
+  margin-top: 20px;
+`
+
+export default function Subscription() {
+    const history = useHistory()
+
+    const goToHome = () => {
+      history.push("/")
+    }
+    return(
+        <RegisterContainer>
+            <button onClick={goToHome}>voltar para a home</button>
+            <h1>Inscrever</h1>
+            <ContainerInputs>
+              <input type="text" placeholder="Nome"/>
+              <input type="text" placeholder="idade"/>
+              <input type="text" placeholder="motivo"/>
+              <input type="text" placeholder="profissão"/>
+              <input type="text" placeholder="país"/>
+            </ContainerInputs>
+            <button>Enviar</button>
+        </RegisterContainer>
+    )
+}
