--- quinzena9/labex/src/styles/AppStyle.js
@@ -1,47 +1,0 @@
-import styled from 'styled-components'
-
-export const Title = styled.header`
-text-align: center;
-border-radius: 1px solid black;
-color:#FDFDFD;
-font-family: 'Source Code Pro', monospace;
-font-weight: 400;
-font-size: 3rem;
-margin-top: 1rem;
-`
-
-export const Header = styled.div`
-display: flex;
-justify-content: space-between;
-`
-
-export const Login = styled.p` 
-position: absolute;
-right: 25px;
-background-color: white;
-border-radius: 100%;
-padding:-5px;
-`
-
-export const ImageLogin = styled.img`
-width: 200px;
-height: 200px;
-border-radius: 100%;
-`
-
-export const Search = styled.p`
-width:200px;
-height:20px;
-`
-
-export const ContainerTravels = styled.div`
-display:flex;
-flex-direction: row;
-align-items: center;
-justify-content: center;
-flex-wrap: wrap;
-
-width:900px;
-
-margin-left: 150px;
-`
