+++ quinzena9/labex/src/styles/CandidateListStyle.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+export const RegisterContainer = styled.div`
+  border-radius: 50px;
+  background-color: white;
+  width: 250px;
+  height: 200px;
+  margin: 50px 10px;
+`
+
+export const ContainerInputs = styled.div`
+  text-align: center;
+  margin-top: 20px;
+`
