+++ quinzena9/labex/src/styles/HomeStyle.js
@@ -1,0 +1,33 @@
+import styled from 'styled-components'
+
+export const Title = styled.header`
+text-align: center;
+border-radius: 1px solid black;
+color:#FDFDFD;
+font-family: 'Source Code Pro', monospace;
+font-weight: 400;
+font-size: 3rem;
+margin-top: 1rem;
+`
+
+export const Header = styled.div`
+display: flex;
+justify-content: space-between;
+`
+
+export const Search = styled.p`
+width:200px;
+height:20px;
+`
+
+export const Container = styled.div`
+display:flex;
+flex-direction: row;
+align-items: center;
+justify-content: center;
+flex-wrap: wrap;
+
+width:900px;
+
+margin-left: 150px;
+`
