+++ quinzena9/labex/src/styles/ListTravelsStyle.js
@@ -1,0 +1,46 @@
+import styled from 'styled-components'
+
+export const Travels = styled.div`
+  border-radius: 50px;
+  background-color: white;
+  width: 250px;
+  height: 400px;
+  margin: 50px 10px;
+` 
+
+export const Display = styled.div`
+  display: flex;
+  justify-content: space-between;
+  text-align: right;
+  font-size: 12px;
+  margin:10px;
+`
+export const Container = styled.div`
+display:flex;
+flex-direction: row;
+align-items: center;
+justify-content: center;
+flex-wrap: wrap;
+
+width:900px;
+
+margin-left: 30px;
+`
+
+export const Login = styled.button` 
+position: absolute;
+top: 5px;
+right: 25px;
+background-color: #0E0127;
+border-radius: 100%;
+border: none;
+outline: thin dotted;
+outline: 0px auto -webkit-focus-ring-color;
+outline-offset: 0px;
+`
+
+export const ImageLogin = styled.img`
+width: 200px;
+height: 200px;
+border-radius: 100%;
+`
