+++ quinzena9/labex/src/styles/RegisterStyle.js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 export const RegisterContainer = styled.div`
   border-radius: 50px;
   background-color: white;
-  width: 250px;
+  width: 300px;
   height: 200px;
   margin: 50px 10px;
 `
