+++ quinzena9/labex/src/styles/style.css
@@ -1,7 +1,7 @@
 body {
     margin: 0;
     padding: 0;
-    background-color: #081960;
+    background-color: #0E0127;
     color: #081960;
   }
